+++ bin/try_doc.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cd doc
+bundle exec jekyll serve
+cd ..