+++ flutter_sound/lib/public/util/wave_header.dart
@@ -155,12 +155,17 @@ class WaveHeader {
     writeId(out, 'WAVE'); // RIFF Form Type
     /* fmt chunk */
     writeId(out, 'fmt ');
-    writeInt(out, 16); // Size of the rest of the Subchunk which follows this number. // 18???
+    writeInt(out,
+        16); // Size of the rest of the Subchunk which follows this number. // 18???
     writeint(out, mFormat);
     writeint(out, mNumChannels);
     writeInt(out, mSampleRate);
-    writeInt(out, (mNumChannels * mSampleRate * mBitsPerSample / 8).floor()); // Average Bytes per second
-    writeint(out, (mNumChannels * mBitsPerSample / 8).floor()); // BlocK Align in bytes
+    writeInt(
+        out,
+        (mNumChannels * mSampleRate * mBitsPerSample / 8)
+            .floor()); // Average Bytes per second
+    writeint(out,
+        (mNumChannels * mBitsPerSample / 8).floor()); // BlocK Align in bytes
     writeint(out, mBitsPerSample);
     /* data chunk */
     writeId(out, 'data');